enabled abilty to use  Node_env to lauch just the backend with no webserver front end. Set to allow larger number set uint32.

+++ src/dataKeys.js
@@ -47,7 +47,7 @@ export const DATA_MAP = {
   BATT_VOLTAGE: { id: keygen(), byteType: TYPES.FLOAT }, // xx.x volts
 
   // Data from GPS
-  ODOMETER:{ id: keygen(), byteType: TYPES.INT16 },// Current Miles Odometer
+  ODOMETER:{ id: keygen(), byteType: TYPES.UINT32 },// Current Miles Odometer
   TRIP_ODOMETER: { id: keygen(), byteType: TYPES.INT16 }, //
   GPS_SPEEED: { id: keygen(), byteType: TYPES.INT16 }, // Speed MPH
